Update 0_analysis_script.R
line 21: removes "mixmgfa" from the list of imported packages (as the package is sourced lines 27-30)
lines 117-126: computes a loneliness score for a given scale only if the participant answered to all items of that scale. Previously, sum scores were calculated even if the participant did not answer to all items (which would result in downwardly biased sum scores)
lines 130-136: checks the number of NAs for age, gender, educational attainment, djgls-6, t-ils, and direct loneliness (we report that information in the manuscript)
lines 140-146 & 167-169: computes the percentage of male/female/other (we report that information in the manuscript)
line 182: save changes in the dataframe
lines 447-457: rename functions/parameters/objects to match those in the mixmgfa package
lines 467-473: changed == 1 to >= .5; Not all countries had 1 values, which means that some countries were not linked to any cluster!
lines 492-506: set the ordered=TRUE parameter (if not done, the scale is considered as continuous, which we don't want as it is 3-point Likert type scale)
lines 527-579: removed the "differiental chisquare" parameter from the values we export, as we don't use it for inferential decisions. Instead, added the cfi/rmsea metrics (used for inferential decisions)
lines 771-816: added ifelse condition to define the estimator method (MLR/WLSMV) and ordered parameter (TRUE/FALSE) based on the measure passed to the cfa() function. This is relevant as the social support measure is considered continuous (estimator=MLR/ordered=FALSE) whereas the loneliness measures are considered ordinal (estimator=WLSMV/ordered=TRUE)
lines 829-832; line 951; lines 1076-1080; line 1136; lines 1163-1072; lines 1192-1195: reordered the variables to match the order in which we call them in the manuscript (in answer to some reviewers' comments)
line 887; line 1173; line 1196: computes correlations using pairwise.complete.obs instead of complete.obs. Correlations are thus computed using much bigger samples.
line 959: changed p > .10 to p >= .10, which matches the "at least .10" wording we use in the manuscript.
lines 1013-1032: changed "proportion > 2/3" to "proportion >= 2/3", which matches the "at least two thirds" wording we use in the manuscript.